+++ client/components/SearchIcon.tsx
