+++ client/helpers/highlight.ts
