+++ client/models/snippets.ts
@@ -1,4 +1,4 @@
-import data from '../data/data.json';
+import data from '../../data/data.json';
 
 interface Snippet {
     body: string;
